Refactor: Move openAPI to backend and CORS to own middleware
wip
