+++ config/backend.go
@@ -17,7 +17,7 @@ type Backend struct {
 	HTTP2                  bool       `hcl:"http2,optional"`
 	MaxConnections         int        `hcl:"max_connections,optional"`
 	Name                   string     `hcl:"name,label"`
-	OpenAPI                []*OpenAPI `hcl:"openapi,block"`
+	OpenAPI                *OpenAPI `hcl:"openapi,block"`
 	PathPrefix             string     `hcl:"path_prefix,optional"`
 	Proxy                  string     `hcl:"proxy,optional"`
 	Remain                 hcl.Body   `hcl:",remain"`
