+++ handler/middleware/next.go
@@ -1,0 +1,7 @@
+package middleware
+
+import "net/http"
+
+type NextHandler interface {
+	ServeNextHTTP(http.ResponseWriter, http.Handler, *http.Request)
+}
